requirements.txt
# Core dependencies
# NOTE: Python 3.14 has compatibility issues with spaCy. Use Python 3.13 or 3.12 for best compatibility.
torch>=2.0.0
nltk>=3.8.0
pdfplumber>=0.9.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
neo4j>=5.0.0
spacy==3.8.7
yake>=0.4.0
python-dateutil>=2.8.0

# Evaluation framework
deepeval>=0.17.0
pytest>=7.0.0

# spaCy English model
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
groq
sentence-transformers
chromadb
langchain_core
langchain_text_splitters
langchain_community
pypdf
tiktoken
pydantic
